Refactor ValidatorRepository by adding an interface
